LPS-58633 refactoring - splitting multi and single resource handling.
